requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
cachetools==5.3.2
deprecation==2.1.0
frozenlist==1.4.1
grpcio==1.60.1
idna==3.6
multidict==6.0.5
packaging==23.2
protobuf==4.25.3
pydantic==2.6.1
pydantic_core==2.16.2
python-dateutil==2.8.2
python-dotenv==1.0.1
six==1.16.0
tinkoff==0.1.1
tinkoff-investments==0.2.0b66
typing_extensions==4.9.0
yarl==1.9.4
